fix: improve typings and organization
