+++ package.json
@@ -22,10 +22,10 @@
     "yarn": "1.22.17"
   },
   "dependencies": {
-    "@changesets/cli": "^2.18.1",
-    "@changesets/changelog-github": "^0.4.2"
+    "@changesets/changelog-github": "^0.4.2",
+    "@changesets/cli": "^2.18.1"
   },
   "devDependencies": {
-    "prettier": "^2.5.0"
+    "prettier": "^2.8.4"
   }
 }
