+++ packages/core/src/components.mts
@@ -1,2 +1,2 @@
-// @ts-expect-error - missing *.astro module typings
+// @ts-ignore - TODO: Fix error
 export { default as Icon } from '../lib/Icon.astro'
