--- packages/core/src/types.d.mts
@@ -1,9 +1,0 @@
-
-
-export type IntegrationOptions = {
-  include?: Record<string, ['*'] | string[]>
-  /**
-   * @default "src/icons"
-   */
-  iconDir?: string
-}
