+++ packages/core/typings/global.d.ts
