+++ packages/core/typings/iconify.d.ts
@@ -1,0 +1,8 @@
+import type { runSVGO } from '@iconify/tools'
+
+export type SVGOOptions = Parameters<typeof runSVGO>[1]
+
+export type { IconifyJSON } from '@iconify/types'
+export type { SVG } from '@iconify/tools'
+export type { Color } from "@iconify/utils/lib/colors/types";
+export type { AutoInstall } from '@iconify/utils/lib/loader/types.js';
